+++ src/App.js
@@ -2,6 +2,8 @@ import { BrowserRouter, Switch, Route } from "react-router-dom";
 
 import Home from "./pages/Home";
 import Page2 from "./pages/Page2";
+import ChooseMajor from "./pages/ChooseMajor";
+
 import "./App.module.css";
 
 const App = () => (
@@ -10,8 +12,11 @@ const App = () => (
       <Route path="/page2">
         <Page2 />
       </Route>
-      <Route path="/">
+      <Route exact path="/">
         <Home />
+      </Route>
+      <Route path="/choose-major">
+        <ChooseMajor />
       </Route>
     </Switch>
   </BrowserRouter>
