+++ src/components/ChooseMajor/ChooseMajor.module.css
@@ -1,0 +1,7 @@
+.subject-box {
+  display: -webkit-flex;
+  display: -moz-flex;
+  display: -ms-flex;
+  display: -o-flex;
+  display: flex;
+}
