+++ src/components/ChooseMajor/MajorSelectList.js
